WIP FuelTank Part
Fuel tank is implemented with State and Proxy patterns.

Could be done easily considering the static limited number of states the tank can be in, but since the project itself is for testing patterns this will work.

FuelState is a Proxy to the actual `quantity` of fuel in the state and works with the FuelTank `capacity` to determine the methods output.
The abstract FuelState class implements factory methods to build the actual implementations