Add fuel tanks (#2)
* add PartsJSON interface

parts toJSON() return value now has a proper interface

* WIP FuelTank Part

Fuel tank is implemented with State and Proxy patterns.

Could be done easily considering the static limited number of states the tank can be in, but since the project itself is for testing patterns this will work.

FuelState is a Proxy to the actual `quantity` of fuel in the state and works with the FuelTank `capacity` to determine the methods output.
The abstract FuelState class implements factory methods to build the actual implementations

* add missing PartJSON imports

required for proper compilation

* fuel tank comments, behaviours and stuff

* added comments and documentation
* added utility methods
* added a new behaviour to the tank: can have infinite fuel
* added a 'standard' implementation at the end for posterity

* Add FuelTanks factory

* add FuelTanks to PartsVendor

vendors can now provide fuel tanks

* add fuel mass

fuel and fuel tanks now have a mass calculated on the fly based on the residual fuel present in the tank

* parts vendor and factories refactor

refactor passed options. For parts that actually have options (for now fuel tanks) we expect a proper object. for anything else we don't care.
Vendors actually pass the options down the line

add exports for FuelTank in index

* add fuel tanks and default options to entitiy factories

we now support creating entities with fuel tanks

* entities major refactor

move a bunch of properties and methods from Ship and Aircraft into the base Vehicle class

add fuel tanks
vehicles now accepts options for the fuel tank as separate option (we may change the interface soon)

Vehicle total mass is now calculated including the fuel tank mass