Targeting & Moving actions as Strategies (#4)
* TargetingSystem

add targeting system with two strategies: one that allows selecting and handling targets, one as NullObject

* entities refactor for targeting system

all methods that handled target management are now moved into the base Entity class and are just proxies for the actual system

* set fishing ship to null targeting

fishing ships are civilians. cannot target enemies.

* expose systems from main module

* add basic moving strategies

* implement moving strategies into entities

to move is now technically an ability of all Entities. default entities cannot move and say so, all other (ships and airplanes) have their own move logic implemented as Strategy

* fix targeting log

* include type constants into base classes .TYPE

this allow to import just the abstract base class to define what something is and use the same import for discriminating on creation.

This change could come handy when the project is compiled.

subclasses are still hidden in the module

* Add single Strategy export object

strategies are now accessible by default export Strategy